move minAbove debug to file with markCandidates
